Refactorizacion de las tarjetas de las materias, asi como la eliminacion de etiquetas como 'Python' y 'Power BI', ademas se elimina una tarjeta de total de presentaciones con codigo
