+++ frontend/src/app/components/filteredSubject.ts
