+++ frontend/src/app/components/subjectCard.types.ts
@@ -2,6 +2,5 @@ import { Module } from "@/types/types";
 
 export interface PresentationCardProps {
   nameSubject: string;
-  githubPath: string;
   modulos: Module[];
 }
