+++ frontend/src/data/subjects-modules.json
@@ -5,7 +5,6 @@
             "nombre": "Matemáticas Avanzadas",
             "codigo": "MAT-2024",
             "descripcion": "Curso de matemáticas para ingeniería",
-            "githubPath": "/matematicas-avanzadas",
             "modulos": [
                 {
                     "id": 1,
@@ -16,13 +15,13 @@
                             "id": 1,
                             "nombre": "Límites y Continuidad",
                             "descripcion": "Introducción al cálculo diferencial",
-                            "githubPath": "/limites-continuidad"
+                            "tipo": "Quarto"
                         },
                         {
                             "id": 2,
                             "nombre": "Derivadas",
                             "descripcion": "Teoría y aplicaciones de derivadas",
-                            "githubPath": "/derivadas"
+                            "tipo": "Quarto"
                         }
                     ]
                 },
@@ -35,13 +34,13 @@
                             "id": 1,
                             "nombre": "Integrales Definidas",
                             "descripcion": "Teoría y aplicaciones de integrales definidas",
-                            "githubPath": "/integrales-definidas"
+                            "tipo": "Quarto"
                         },
                         {
                             "id": 2,
                             "nombre": "Ecuaciones Diferenciales",
                             "descripcion": "Teoría y aplicaciones de ecuaciones diferenciales",
-                            "githubPath": "/ecuaciones-diferenciales"
+                            "tipo": "Quarto"
                         }
                     ]
                 }
@@ -52,7 +51,6 @@
             "nombre": "Programación Web",
             "codigo": "WEB-2024",
             "descripcion": "Desarrollo de aplicaciones web modernas",
-            "githubPath": "/programacion-web",
             "modulos": [
                 {
                     "id": 1,
@@ -63,13 +61,13 @@
                             "id": 1,
                             "nombre": "Estructura HTML",
                             "descripcion": "Elementos básicos de HTML",
-                            "githubPath": "/estructura-html"
+                            "tipo": "Quarto"
                         },
                         {
                             "id": 2,
                             "nombre": "Estilos con CSS",
                             "descripcion": "Diseño y estilos con CSS",
-                            "githubPath": "/estilos-css"
+                            "tipo": "Quarto"
                         }
                     ]
                 },
@@ -82,13 +80,13 @@
                             "id": 1,
                             "nombre": "Funciones y Objetos",
                             "descripcion": "Conceptos avanzados de JavaScript",
-                            "githubPath": "/funciones-objetos"
+                            "tipo": "Quarto"
                         },
                         {
                             "id": 2,
                             "nombre": "Asincronía",
                             "descripcion": "Manejo de operaciones asincrónicas",
-                            "githubPath": "/asincronia"
+                            "tipo": "Quarto"
                         }
                     ]
                 }
@@ -99,7 +97,6 @@
             "nombre": "Análisis de Datos",
             "codigo": "DATA-2024",
             "descripcion": "Fundamentos de ciencia de datos",
-            "githubPath": "/analisis-datos",
             "modulos": [
                 {
                     "id": 1,
@@ -110,7 +107,7 @@
                             "id": 1,
                             "nombre": "Pandas Básico",
                             "descripcion": "Manipulación de datos con Pandas",
-                            "githubPath": "/pandas-basico"
+                            "tipo": "Quarto"
                         }
                     ]
                 },
@@ -123,7 +120,7 @@
                             "id": 1,
                             "nombre": "Gráficos con Matplotlib",
                             "descripcion": "Creación de gráficos en Python",
-                            "githubPath": "/graficos-matplotlib"
+                            "tipo": "Quarto"
                         }
                     ]
                 }
@@ -134,7 +131,6 @@
             "nombre": "Machine Learning",
             "codigo": "ML-2024",
             "descripcion": "Aprendizaje automático aplicado",
-            "githubPath": "/machine-learning",
             "modulos": [
                 {
                     "id": 1,
@@ -145,7 +141,7 @@
                             "id": 1,
                             "nombre": "Regresión Lineal",
                             "descripcion": "Modelo de regresión lineal",
-                            "githubPath": "/regresion-lineal"
+                            "tipo": "Quarto"
                         }
                     ]
                 },
@@ -158,7 +154,7 @@
                             "id": 1,
                             "nombre": "Perceptrón Multicapa",
                             "descripcion": "Fundamentos de redes neuronales",
-                            "githubPath": "/perceptron-multicapa"
+                            "tipo": "Quarto"
                         }
                     ]
                 }
